requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Dance==7.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
SQLAlchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
requests==2.32.5
gunicorn==21.2.0

# Database drivers
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0
factory-boy==3.3.0
pytest-flask==1.3.0
responses==0.24.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Development
python-dotenv==1.0.0
